requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
blas=2.116=mkl
blas-devel=3.9.0=16_linux64_mkl
brotli-python=1.1.0=py39h3d6467e_1
bzip2=1.0.8=hd590300_5
ca-certificates=2024.2.2=hbcca054_0
certifi=2024.2.2=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
datasets=2.19.0=pypi_0
decorator=5.1.1=pypi_0
dill=0.3.8=pypi_0
docker-pycreds=0.4.0=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.14.0=pyhd8ed1ab_0
freetype=2.12.1=h267a509_2
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gmp=6.3.0=h59595ed_1
gmpy2=2.1.5=py39h03b5d36_0
gnutls=3.6.13=h85f3911_1
huggingface-hub=0.23.0=pypi_0
icu=73.2=h59595ed_0
idna=3.7=pyhd8ed1ab_0
ipython=8.18.1=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pyhd8ed1ab_0
jpeg=9e=h166bdaf_2
lame=3.100=h166bdaf_1003
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.40=h55db66e_0
lerc=4.0.0=h27087fc_0
libblas=3.9.0=16_linux64_mkl
libcblas=3.9.0=16_linux64_mkl
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h0b41bf4_0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h77fa898_6
libgfortran-ng=13.2.0=h69a702a_6
libgfortran5=13.2.0=h43f5ff8_6
libgomp=13.2.0=h77fa898_6
libhwloc=2.10.0=default_h2fb2949_1000
libiconv=1.17=hd590300_2
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=16_linux64_mkl
liblapacke=3.9.0=16_linux64_mkl
libnpp=11.8.0.86=0
libnsl=2.0.1=hd590300_0
libnvjpeg=11.9.0.86=0
libpng=1.6.43=h2797004_0
libsqlite=3.45.3=h2797004_0
libstdcxx-ng=13.2.0=hc0a3c3a_6
libtiff=4.5.0=h6adf6a1_2
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.13=h7f98852_1004
libxcrypt=4.4.36=hd590300_1
libxml2=2.12.6=h232c23b_2
libzlib=1.2.13=hd590300_5
lightning-utilities=0.11.2=pypi_0
llvm-openmp=15.0.7=h0cdce71_0
markupsafe=2.1.5=py39hd1e30aa_0
matplotlib-inline=0.1.7=pypi_0
mkl=2022.1.0=h84fe81f_915
mkl-devel=2022.1.0=ha770c72_916
mkl-include=2022.1.0=h84fe81f_915
mpc=1.3.1=hfe3b2da_0
mpfr=4.2.1=h9458935_1
mpmath=1.3.0=pyhd8ed1ab_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4.20240210=h59595ed_0
nettle=3.6=he412f7d_0
networkx=3.2.1=pyhd8ed1ab_0
numpy=1.26.4=py39h474f0d3_0
openh264=2.1.1=h780b84a_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.3.0=hd590300_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=9.4.0=py39h2320bf1_1
pip=24.0=pyhd8ed1ab_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=16.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pygments=2.18.0=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.19=h0755675_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.9=4_cp39
pytorch=2.3.0=py3.9_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-lightning=2.2.4=pypi_0
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py39hd1e30aa_1
readline=8.2=h8228510_1
regex=2024.4.28=pypi_0
requests=2.31.0=pyhd8ed1ab_0
safetensors=0.4.3=pypi_0
sentry-sdk=2.0.1=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.5.1=pyhd8ed1ab_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
stack-data=0.6.3=pypi_0
sympy=1.12=pypyh9d50eac_103
tbb=2021.12.0=h00ab1b0_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.19.1=pypi_0
torchaudio=2.3.0=py39_cu118
torchmetrics=1.3.2=pypi_0
torchsummary=1.5.1=pypi_0
torchtriton=2.3.0=py39
torchvision=0.18.0=py39_cu118
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.40.1=pypi_0
typing_extensions=4.11.0=pyha770c72_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pyhd8ed1ab_0
wandb=0.16.6=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xxhash=3.4.1=pypi_0
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
yarl=1.9.4=pypi_0
zlib=1.2.13=hd590300_5
zstd=1.5.6=ha6fb4c9_0